Create the Watchlist::Item model
It creates the `Watchlist::Item` model, which is going to replace the
`WatchedProject` model in the users' watchlists context.

The `Watchlist::Item` model features a polymorphic `watchable`
association which makes it usable for associating records of any type,
whereas the `WatchedProject` model can be used for `Project`
records only.

The migration takes care of creating the `watchlist_items` table and of
populating it basing on the existing `watched_projects` records. It's
written in SQL because it should be considerably faster than the
standard ORM implementation.